dsvs

+++ src/domain/arenas/ArenaDaoImpl.java
@@ -17,37 +17,63 @@ public class ArenaDaoImpl implements ArenaDao{
 	DbManager db = new DbManager();
 	
 	@Override
-	public void addArena(Arena a) {
+	public int addArena(Arena a) {
 		int status = 0;
 		try{
 			conn = db.getConnection();
-			ps =conn.prepareStatement("insert into arenas (name,location) values(?,?)");
-			//ps.setString(1, a.setarenaID(arenaID));
-			ps.setString(1, a.getLocation());
-			ps.setString(2, a.getName());
-			status = ps.executeUpdate();
-			conn.close();
-		}catch(Exception e){
-			System.out.println(e);
-		}
-		//return status;
-	}
-	@Override
-	public void deleteArena(Arena a) {
-		int status = 0;
-		try{
-			conn = db.getConnection();
-			ps =conn.prepareStatement("delete from arenas where name = ? and location = ?");
-			//ps.setString(1, a.setarenaID(arenaID));
-			System.out.println(a.getLocation() +" "+ a.getName());
+			ps =conn.prepareStatement("Select * from arenas where name=? and location=?");
 			ps.setString(2, a.getLocation());
 			ps.setString(1, a.getName());
-			status = ps.executeUpdate();
+			ResultSet rs = ps.executeQuery();
+			while(rs.next())
+			{
+				status = 1;
+			}
+			if(status==0)
+			{
+				ps =conn.prepareStatement("insert into arenas (name,location) values(?,?)");
+				//ps.setString(1, a.setarenaID(arenaID));
+				ps.setString(2, a.getLocation());
+				ps.setString(1, a.getName());
+				status = ps.executeUpdate();
+				System.out.println(status);
+				status = 0;
+			}
+			
 			conn.close();
 		}catch(Exception e){
 			System.out.println(e);
 		}
-		//return status;
+		return status;
+	}
+	@Override
+	public int deleteArena(Arena a) {
+		int status = 0;
+		try{
+			conn = db.getConnection();
+			ps =conn.prepareStatement("Select * from arenas where name=? and location=?");
+			ps.setString(2, a.getLocation());
+			ps.setString(1, a.getName());
+			ResultSet rs = ps.executeQuery();
+			while(rs.next())
+			{
+				status = 1;
+			}
+			if(status==1)
+			{
+				ps =conn.prepareStatement("delete from arenas where name = ? and location = ?");
+				//ps.setString(1, a.setarenaID(arenaID));
+				System.out.println(a.getLocation() +" "+ a.getName());
+				ps.setString(2, a.getLocation());
+				ps.setString(1, a.getName());
+				status = ps.executeUpdate();
+			}
+			
+			conn.close();
+		}catch(Exception e){
+			System.out.println(e);
+		}
+		return status;
 	}
 	@Override
 	public List<Arena> displayArenas() {
